requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.128.0
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.81
langchain-openai==0.2.14
langchain-text-splitters==0.3.11
langsmith==0.4.34
marshmallow==3.26.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.0
openai==1.109.1
orjson==3.11.3
packaging==25.0
pgvector==0.4.2
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.0
pydantic-settings==2.12.0
pydantic_core==2.41.1
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.40.0
yarl==1.22.0
zstandard==0.25.0
